requirements.txt
torch
transformers
datasets
accelerate
numpy
scipy
scikit-learn
pandas
matplotlib
seaborn
pyyaml
tqdm
umap-learn
safetensors